Move helper functions over to utils
